Clear the full /W4 warning backlog: unused variables, dead code, real bugs
Both Debug and Release now build with zero warnings (was ~280). Three
categories of change:

Mechanical (majority of warning volume): missing 'f' suffixes on float
literals in three large lookup tables (anorms.h, anorm_dots.h,
gl_warp_sin.h) caused ~195 double-to-float truncation warnings on their
own -- fixed via a scripted regex pass. The remaining scattered
truncation/signed-unsigned-mismatch/size_t-to-int warnings got explicit
casts at the specific call site rather than widening shared types like
viddef_t's width/height (unsigned, inconsistent with every other
width/height in the codebase) -- narrower blast radius than retyping a
pervasively-used struct just to silence a handful of comparisons.

Unused variables (~80 across ~20 files, gl_draw.cpp/gl_mesh.cpp/
gl_model.cpp/gl_rmain.cpp/gl_vidnt.cpp had the most): removed dead
declarations, going beyond what the compiler flagged in a few spots
(e.g. gl_vidnt.cpp's VID_MenuDraw had a `dup` int the compiler never
warned about, sitting next to the `dupmode` it did flag). Also removed
a `#if 0`-disabled block in Draw_Init and a DumpChunks() function in
snd_mem.cpp whose only remaining reference was inside a comment --
both had become entirely dead weight once their unused locals were
traced back to their only callers (none).

Real bugs found while tracing "unused" variables back to their actual
callers, not just papered over:
- gl_model.cpp: Mod_LoadAliasFrame/Mod_LoadAliasGroup both had a
  copy-paste bug writing bboxmax source data into the bboxmin
  destination twice, leaving frame->bboxmax permanently unpopulated.
  Currently inert (nothing reads these fields), but fixed correctly
  since nothing else needs to change for it to become live again.
- gl_mesh.cpp: BuildTris' `used[]` array relies on 3 distinct states
  (0/1/2) but was declared qboolean (=bool in this codebase) -- the
  C++ port silently collapsed the "2" sentinel into "1" (true),
  permanently breaking the temp-mark cleanup in StripLength/FanLength
  within a single BuildTris() call. Only affects first-time (cache-miss)
  alias model mesh optimization; fixed by widening used[] to byte with
  a comment explaining why it can't be qboolean.
- in_win.cpp: RawValuePointer's switch had no default case, so an
  out-of-range axis index would return an uninitialized garbage
  pointer (never triggered today -- its only caller passes exactly the
  6 valid indices -- but the compiler was right to flag the fallthrough).
  Added a default case; also marked Sys_Error [[noreturn]] since every
  code path in it ends in exit(1), which was needed for the compiler to
  recognize the default case's Sys_Error call as terminal.
- sys_win.cpp: deleted the GetVersionEx-based platform check entirely
  (deprecated API, C4996) rather than just swapping in a modern
  replacement -- the WinNT global it computed was never read anywhere,
  and the version floor it enforced (Win95/NT4+) can't meaningfully
  fail on any system capable of running a build that targets the
  Windows 10 SDK.

Verified via a clean Debug+Release rebuild (0 warnings, 0 errors, both
configs) and a -condebug smoke test across all 6 main episode levels
(e1m1-e1m4, e1m7, e1m8) -- wpoly/epoly counts match previously-recorded
baselines exactly, WASAPI/XInput/raw-input all initialize cleanly, zero
Sys_Error across the whole run.

+++ WinQuake/cl_demo.cpp
@@ -118,14 +118,14 @@ int CL_GetMessage (void)
 		VectorCopy (cl.mviewangles[0], cl.mviewangles[1]);
 		for (i=0 ; i<3 ; i++)
 		{
-			r = fread (&f, 4, 1, cls.demofile);
+			r = (int)fread (&f, 4, 1, cls.demofile);
 			cl.mviewangles[0][i] = LittleFloat (f);
 		}
-		
+
 		net.message.cursize = LittleLong (net.message.cursize);
 		if (net.message.cursize > MAX_MSGLEN)
 			Sys_Error ("Demo message > MAX_MSGLEN");
-		r = fread (net.message.data, net.message.cursize, 1, cls.demofile);
+		r = (int)fread (net.message.data, net.message.cursize, 1, cls.demofile);
 		if (r != 1)
 		{
 			CL_StopPlayback ();
